requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
requests
aiofiles
jinja2
sqlalchemy
pydantic
pydantic-settings
httpx